+++ src/Kiota.Builder/Refiners/ILanguageRefiner.cs
@@ -32,9 +32,6 @@ public interface ILanguageRefiner
             case GenerationLanguage.Go:
                 await new GoRefiner(config).RefineAsync(generatedCode, cancellationToken).ConfigureAwait(false);
                 break;
-            case GenerationLanguage.Swift:
-                await new SwiftRefiner(config).RefineAsync(generatedCode, cancellationToken).ConfigureAwait(false);
-                break;
             case GenerationLanguage.HTTP:
                 await new HttpRefiner(config).RefineAsync(generatedCode, cancellationToken).ConfigureAwait(false);
                 break;