+++ src/Kiota.Builder/Writers/LanguageWriter.cs
@@ -13,7 +13,6 @@ using Kiota.Builder.Writers.Java;
 using Kiota.Builder.Writers.Php;
 using Kiota.Builder.Writers.Python;
 using Kiota.Builder.Writers.Ruby;
-using Kiota.Builder.Writers.Swift;
 using Kiota.Builder.Writers.TypeScript;
 
 namespace Kiota.Builder.Writers;
@@ -190,7 +189,6 @@ public abstract class LanguageWriter
             GenerationLanguage.PHP => new PhpWriter(outputPath, clientNamespaceName, usesBackingStore),
             GenerationLanguage.Python => new PythonWriter(outputPath, clientNamespaceName, usesBackingStore),
             GenerationLanguage.Go => new GoWriter(outputPath, clientNamespaceName, excludeBackwardCompatible),
-            GenerationLanguage.Swift => new SwiftWriter(outputPath, clientNamespaceName),
             GenerationLanguage.Dart => new DartWriter(outputPath, clientNamespaceName),
             GenerationLanguage.HTTP => new HttpWriter(outputPath, clientNamespaceName),
             _ => throw new InvalidEnumArgumentException($"{language} language currently not supported."),