+++ vscode/packages/microsoft-kiota/src/util.ts
@@ -104,8 +104,6 @@ export function parseGenerationLanguage(value: string): KiotaGenerationLanguage 
           return KiotaGenerationLanguage.Python;
       case "go":
           return KiotaGenerationLanguage.Go;
-      case "swift":
-          return KiotaGenerationLanguage.Swift;
       case "ruby":
           return KiotaGenerationLanguage.Ruby;
       case "dart":